requirements.txt
absl-py==0.11.0
appdirs==1.4.3
appnope==0.1.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.2.0
biopython==1.78
bleach==3.1.0
cachetools==4.2.0
certifi==2019.9.11
cffi==1.14.3
chardet==3.0.4
click==8.0.1
configparser==4.0.2
cycler==0.10.0
Cython==0.29.13
dataclasses==0.6
DateTime==4.3
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.1
entrypoints==0.3
enum34==1.1.6
filelock==3.0.12
future==0.18.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.1
h5py==2.10.0
html5lib==1.0.1
huggingface-hub==0.0.12
idna==2.10
importlib-metadata==3.1.0
info-gain==1.0.1
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.10.3
joblib==1.0.1
json5==0.9.5
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
Keras==2.4.3
kiwisolver==1.3.1
langdetect==1.0.9
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mpmath==1.2.1
nbclient==0.5.1
nbconvert==5.6.1
nbformat==4.4.0
nest-asyncio==1.4.3
nltk==3.6.2
notebook==6.0.1
numpy==1.17.3
oauthlib==3.1.0
olefile==0.46
packaging==21.0
pandas==0.25.2
pandocfilters==1.4.2
parso==0.7.1
pathlib2==2.3.5
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.9.0
prompt-toolkit==2.0.10
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.17.3
python-dateutil==2.8.0
pytorch-lightning==0.7.6
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
regex==2021.4.4
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.7
sacremoses==0.0.45
scandir==1.10.0
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.9.0
Send2Trash==1.5.0
sentence-transformers==2.0.0
sentencepiece==0.1.96
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
sympy==1.7.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
terminado==0.8.2
testpath==0.4.2
threadpoolctl==2.1.0
tk==0.1.0
tokenizers==0.10.3
torch==1.9.0
torchvision==0.10.0
tornado==6.0.3
tqdm==4.42.0
traitlets==4.3.3
transformers==4.8.2
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.0
zope.interface==5.4.0
